refactor capabilities handling
